[MOFLIS] #16 매칭글 기능 test수정

+++ src/main/java/com/project/moflis/participant/command/ApplyParticipantCommand.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @Setter
 @AllArgsConstructor
 public class ApplyParticipantCommand {
-    private int postId;
+    private Long postId;
     private ParticipantStatus status;
     private LocalDateTime joinDate;
     private LocalDateTime leaveDate;
